requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
streamlit-authenticator
bcrypt
pyyaml
plotly
sqlite3